// src/game/enums/game-status.enum.ts
export enum GameStatus {
  PENDING = 'PENDING',
  READY_TO_START = 'READY_TO_START',
  IN_PROGRESS = 'IN_PROGRESS',
  ROUND_IN_PROGRESS = 'ROUND_IN_PROGRESS',
  ROUND_ENDED = 'ROUND_ENDED',
  WAITING_FOR_TEAMS = 'WAITING_FOR_TEAMS',
  PAUSED = 'PAUSED',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
}